updated_youtube_project/requirements.txt
streamlit
pandas
mysql-connector-python
google-api-python-client
isodate
